:hammer: use RetryIf and only log we died if we're going to restart
